pyproject.toml
[tool.poetry]
name = "moderation"
version = "0.1.0"
description = "Moderation for GenAI"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
modal-client = "^0.48.1853"
openai2 = "^1.5.1"
loguru = "^0.7.0"
streamlit = "^1.21.0"
anthropic = "^0.2.6"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
